Updated for latest game versions, various bug fixes
PSO structure appears to no longer hold the JOAAT hash of the camera metadata class, so we've just removed this check for now which otherwise would cause an early out in checkCameraFrame(). Instead of implicitly calling pattern.get() after an AOB scan, we check that the pattern scan actually succeeded then resolve the result based on that. Fixes some va_args that were passed without being in the format string. Adds more concise naming to some logging and functions. If presets file is deleted while the game is running, we detect that and create a new one before saving new presets.